corrigido o bug do bigdecimal

+++ src/main/java/com/example/banco/model/DadosContaModel.java
@@ -2,15 +2,14 @@ package com.example.banco.model;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 @Data
 public class DadosContaModel {
     private Long id;
     private String agencia;
     private String numero;
     private String titular;
-    private BigDecimal saldoAtual;
-    private BigDecimal saldoFinal;
-    private BigDecimal valorFornecido;
+    private double saldoAtual;
+    private double saldoFinal;
+    private double valorFornecido;
     private String servico;
 }
